+++ components/formFields.styles.js
@@ -1,6 +1,6 @@
 import { StyleSheet } from 'react-native';
 import { COLORS } from '../theme/colors';
-import { SPACINGS, BORDERWIDTH, BORDERRADIUS } from '../theme/spacings';
+import { SPACINGS, BORDERWIDTH } from '../theme/spacings';
 import { SIZES, LINEHEIGHTS, FONTWEIGHT } from '../theme/fonts';
 
 export const styles = StyleSheet.create({
@@ -22,35 +22,8 @@ export const styles = StyleSheet.create({
     height: SPACINGS.multiplyBy(2),
     fontSize: SIZES.md,
   },
-  customCheckBoxTitle: {
-    color: COLORS.text.main
-  },
-  customCheckBoxTitleChecked: {
-    color: COLORS.text.inverted
-  },
   checkBoxContainer: {
     backgroundColor: COLORS.text.inverted,
-  },
-  customCheckBoxButton: {
-    backgroundColor: COLORS.text.inverted,
-    width: '30%',
-    flexDirection: 'row',
-    flexWrap: 'wrap',
-    marginLeft: 0,
-    borderRadius: BORDERRADIUS.lg
-  },
-  customCheckBoxButtonChecked: {
-    color: COLORS.text.inverted,
-    backgroundColor: COLORS.primary,
-    width: '30%',
-    flexDirection: 'row',
-    flexWrap: 'wrap',
-    marginLeft: 0,
-    borderRadius: BORDERRADIUS.lg,
-  },
-  customCheckBoxContainer: {
-    flexDirection: 'row',
-    flexWrap: 'wrap'
   },
   dropdownFormField: {
     borderBottomColor: COLORS.grey.extralight,
